# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Photo Transformer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransformResultStatus =
  (typeof TransformResultStatus)[keyof typeof TransformResultStatus];

export const TransformResultStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface TransformResult {
  /** Job ID for polling status */
  jobId: string;
  status: TransformResultStatus;
  /** URL of the transformed image (when completed) */
  imageUrl?: string;
  /** The preset that was applied */
  preset: string;
}

export type TransformStatusStatus =
  (typeof TransformStatusStatus)[keyof typeof TransformStatusStatus];

export const TransformStatusStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface TransformStatus {
  jobId: string;
  status: TransformStatusStatus;
  /** URL of the transformed image (when completed) */
  imageUrl?: string;
  /** Error message (when failed) */
  error?: string;
  preset: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

/**
 * The cinematic style preset to apply
 */
export type TransformImageBodyPreset =
  (typeof TransformImageBodyPreset)[keyof typeof TransformImageBodyPreset];

export const TransformImageBodyPreset = {
  cinematic: "cinematic",
  scifi: "scifi",
  neo_noir: "neo_noir",
  warm_hollywood: "warm_hollywood",
  dramatic_portrait: "dramatic_portrait",
} as const;

export type TransformImageBody = {
  /** The image file to transform (JPG or PNG) */
  image: Blob;
  /** The cinematic style preset to apply */
  preset: TransformImageBodyPreset;
  /** Whether to add letterbox bars (2.39:1 aspect ratio) */
  letterbox?: boolean;
};