# lib/api-zod/src/generated/types/transformResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Photo Transformer API
 * OpenAPI spec version: 0.1.0
 */
import type { TransformResultStatus } from "./transformResultStatus";

export interface TransformResult {
  /** Job ID for polling status */
  jobId: string;
  status: TransformResultStatus;
  /** URL of the transformed image (when completed) */
  imageUrl?: string;
  /** The preset that was applied */
  preset: string;
}