# lib/api-zod/src/generated/types/transformStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cinematic AI Photo Transformer API
 * OpenAPI spec version: 0.1.0
 */
import type { TransformStatusStatus } from "./transformStatusStatus";

export interface TransformStatus {
  jobId: string;
  status: TransformStatusStatus;
  /** URL of the transformed image (when completed) */
  imageUrl?: string;
  /** Error message (when failed) */
  error?: string;
  preset: string;
}